몬스터 공격 구현, 아이템 중복 제거, 총알 변경
몬스터의 공격 판정, 플레이어 피해 구현
아이템 중복 출현 제거
총알 변경

+++ Assets/Scripts/Inventory.cs
@@ -39,7 +39,7 @@ public class Inventory : MonoBehaviour
                 PlayerValue.Instance.IncreaseAttackSpeed(item.itemStat);
                 break;
             case "치명타 확률업":
-                PlayerValue.Instance.IncreaseCritChance(item.itemStat);
+                PlayerValue.Instance.IncreaseCritChance((int)item.itemStat);
                 break;
             case "이동속도 템":
                 PlayerValue.Instance.IncreaseMoveSpeed(item.itemStat);
